+++ app/src/main/java/com/juan/pinya/fragments/dailyreport/DRAdd1Adapter.kt
@@ -10,25 +10,25 @@ import com.juan.pinya.R
 import com.juan.pinya.model.DailyReport
 import kotlinx.android.synthetic.main.item_addlist.view.*
 
-class DRA2Adapter(options: FirestoreRecyclerOptions<DailyReport>,
-                  private val listener: DailyReportAdd2Fragment
+class DRAdd1Adapter(options: FirestoreRecyclerOptions<DailyReport>,
+                    private val listener: DRAdd1Fragment
 ) :
-    FirestoreRecyclerAdapter<DailyReport, DRA2Adapter.DRA2ViewHoder>(options) {
+    FirestoreRecyclerAdapter<DailyReport, DRAdd1Adapter.DRA1ViewHoder>(options) {
 
-    class DRA2ViewHoder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    class DRA1ViewHoder(itemView: View) : RecyclerView.ViewHolder(itemView) {
 
         var company = itemView.add_company_textView
 
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): DRA2ViewHoder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): DRA1ViewHoder {
         val itemView =
             LayoutInflater.from(parent.context).inflate(R.layout.item_addlist, parent, false)
-        return DRA2ViewHoder(itemView)
+        return DRA1ViewHoder(itemView)
     }
 
     override fun onBindViewHolder(
-        holder: DRA2ViewHoder,
+        holder: DRA1ViewHoder,
         position: Int,
         dailyReport: DailyReport
     ) {
