+++ app/src/main/res/layout/fragment_daily_report_add2.xml
@@ -6,7 +6,7 @@
     android:background="@color/pinya_yellow2"
     android:gravity="center_horizontal"
     android:orientation="vertical"
-    tools:context=".fragments.dailyreport.DailyReportAdd2Fragment">
+    tools:context=".fragments.dailyreport.DRAdd2Fragment">
 
     <TextView
         android:id="@+id/add2_textView"
