+++ app/src/main/res/layout/item_addlist.xml
@@ -1,9 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    android:layout_margin="5dp">
+    android:layout_margin="5dp"
+    android:orientation="horizontal">
 
     <TextView
         android:id="@+id/add_company_textView"
@@ -12,5 +14,17 @@
         android:textSize="20dp"
         android:minWidth="100dp"
         android:gravity="center"
+        android:background="@color/pinya_checked"
         tools:text="廠商Ａ" />
-</androidx.cardview.widget.CardView>
+
+    <TextView
+        android:id="@+id/add_piece_textView"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:gravity="center"
+        android:minWidth="100dp"
+        android:textSize="20dp"
+        android:background="@color/pinya_checked"
+        android:visibility="gone"
+        tools:text="180" />
+</LinearLayout>
