+++ app/src/main/res/values/strings.xml
@@ -19,4 +19,12 @@
     <string name="text_next">下一步</string>
     <string name="text_cancel">取消</string>
     <string name="text_up">上一步</string>
+    <string name="text_compost">廠商：</string>
+    <string name="text_site">工地：</string>
+    <string name="text_address">地點：</string>
+    <string name="text_price">單價：</string>
+    <string name="text_meter">米數：</string>
+    <string name="text_date">時間：</string>
+    <string name="text_number">車次：</string>
+    <string name="text_ps">備註：</string>
 </resources>
